预约课程功能

+++ src/main/java/cn/edu/mju/ccce/dtrsystem/bean/Reservation.java
@@ -62,7 +62,7 @@ public class Reservation implements Serializable {
         return USER_NBR;
     }
 
-    public void setUSER_NBR(int USER_NBR) {
+    public void setUSER_NBR(Long USER_NBR) {
         this.USER_NBR = USER_NBR;
     }
